--- src/components/ActionChip/index.ts
@@ -1,1 +1,0 @@
-export { default as ActionChip } from './ActionChip.vue'
