--- src/components/ConfirmDialog/index.ts
@@ -1,1 +1,0 @@
-export { default as ConfirmDialog } from './ConfirmDialog.vue'
