--- src/components/ConfirmDialog/package.json
@@ -1,10 +1,0 @@
-{
-  "name": "@bcrs-shared-components/confirm-dialog",
-  "version": "1.0.0",
-  "publishConfig": {
-    "access": "public"
-  },
-  "dependencies": {
-    "vue-property-decorator": "^8.4.0"
-  }
-}
