+++ src/components/action-chip/index.ts
@@ -1,0 +1,1 @@
+export { default as ActionChip } from './ActionChip.vue'
