+++ src/components/action-chip/package.json
@@ -1,0 +1,12 @@
+{
+  "name": "@bcrs-shared-components/action-chip",
+  "version": "1.0.0",
+  "publishConfig": {
+    "access": "public"
+  },
+  "dependencies": {
+    "@bcrs-shared-components/enums": "^1.0.9",
+    "@bcrs-shared-components/interfaces": "^1.0.14",
+    "vue-property-decorator": "^8.4.0"
+  }
+}
