+++ src/components/confirm-dialog/index.ts
@@ -1,0 +1,1 @@
+export { default as ConfirmDialog } from './ConfirmDialog.vue'
