+++ src/mixins/index.ts
@@ -1,9 +1,7 @@
-import CommonMixin from './common-mixin'
 import DateMixin from './date-mixin'
 import FilingMixin from './filing-mixin'
 
 export {
-  CommonMixin,
   DateMixin,
   FilingMixin
 }
