+++ tests/unit/ActionChip.spec.ts
@@ -2,7 +2,7 @@ import Vue from 'vue'
 import Vuetify from 'vuetify'
 import Affix from 'vue-affix'
 import { shallowMount } from '@vue/test-utils'
-import { ActionChip } from '@/components/ActionChip'
+import { ActionChip } from '@/components/action-chip'
 import { ActionTypes } from '@bcrs-shared-components/enums'
 
 Vue.use(Vuetify)
@@ -11,27 +11,27 @@ Vue.use(Affix)
 const vuetify = new Vuetify({})
 
 describe('ActionChip', () => {
-  const wrapperFactory = (action: ActionTypes, isCorrection: boolean) => {
+  const wrapperFactory = (action: ActionTypes, editedLabel: string = null) => {
     return shallowMount(ActionChip, {
       propsData: {
         actionableItem: {
           action
         },
-        isCorrection
+        editedLabel
       },
       vuetify
     })
   }
 
   it('initializes correctly', () => {
-    const wrapper = wrapperFactory('added', true)
+    const wrapper = wrapperFactory('added')
 
     // verify that component mounts
     expect(wrapper.find(ActionChip).exists()).toBe(true)
   })
 
   it('displays the added chip correctly', () => {
-    const wrapper = wrapperFactory('added', true)
+    const wrapper = wrapperFactory('added')
 
     // verify that component mounts
     expect(wrapper.find(ActionChip).exists()).toBe(true)
@@ -39,7 +39,7 @@ describe('ActionChip', () => {
   })
 
   it('displays the edited chip correctly for a correction', async () => {
-    const wrapper = wrapperFactory('edited', true)
+    const wrapper = wrapperFactory('edited', 'CORRECTED')
 
     // verify that component mounts
     expect(wrapper.find(ActionChip).exists()).toBe(true)
@@ -47,7 +47,7 @@ describe('ActionChip', () => {
   })
 
   it('displays the edited chip correctly for an alteration', () => {
-    const wrapper = wrapperFactory('edited', false)
+    const wrapper = wrapperFactory('edited', 'CHANGED')
 
     // verify that component mounts
     expect(wrapper.find(ActionChip).exists()).toBe(true)
@@ -55,7 +55,7 @@ describe('ActionChip', () => {
   })
 
   it('displays the added chip correctly', () => {
-    const wrapper = wrapperFactory('removed', true)
+    const wrapper = wrapperFactory('removed')
 
     // verify that component mounts
     expect(wrapper.find(ActionChip).exists()).toBe(true)
